Attempting to create resources from PSON
I'm rather loathe to commit this in a non-working state, but I'm going
to be out of town all next week and other people should be able to work
on this. New action puppetral#create_from_pson should be able to create
a resource from the pson data returned by puppetral#find, but...it can't
yet. No real insight as to why, seems like it should be working (since
puppetral#create works and uses much of the same code). Best of luck
getting it working.